+++ test-data/src/main/java/org/ehrbase/test_data/folder/FolderTestDataCanonicalJson.java
@@ -23,7 +23,9 @@ public enum FolderTestDataCanonicalJson {
 
     SIMPLE_EMPTY_FOLDER("A simple empty folder with no sub folders", "simple_empty_folder.json"),
     FLAT_FOLDER_INSERT("A flat folder structure for testing simple inserts.", "flat_folder_insert.json"),
-    NESTED_FOLDER("A folder with muliple nested folders", "nested_folder.json");
+    NESTED_FOLDER("A folder with muliple nested folders", "nested_folder.json"),
+    FOLDER_WITH_DUPLICATE_NAMES("Folder with two folders on the same sibling level with same names", "duplicate_folder_names.json"),
+    FOLDER_WITHOUT_DUPLICATE_NAMES("Folder with two sub folders that have different names", "folder_without_duplicates.json");
 
     private final String description;
     private final String filename;